EZ-1336 fix executor load calculation (2/40==0)

+++ core/src/main/kotlin/core/aas/AutoGradeScheduler.kt
@@ -73,9 +73,8 @@ class AutoGradeScheduler : ApplicationListener<ContextRefreshedEvent> {
 
             getCapableExecutors(autoExerciseId)
                 .mapNotNull { it.associateWithSchedulerOrNull(executors[it.id]?.get(priority)) }
-                .minByOrNull { it.functionScheduler.size() / it.capableExecutor.maxLoad }
+                .minByOrNull { it.functionScheduler.size().toDouble() / it.capableExecutor.maxLoad }
                 ?: throw NoExecutorsException("No capable executors found for this auto exercise")
-
         }
 
         return selectedExecutor.functionScheduler.scheduleAndAwait(selectedExecutor.capableExecutor, request)
